fixed initial data to db fixed logstash level as [long] issue by mutating all level fields to string
