+++ api/main.py
@@ -8,7 +8,6 @@ from api.core.exceptions.handlers import (
     unauthorized_user_handler,
 )
 from api.core.logging.settings import setup_logging
-from api.utils.database import create_super_user
 
 from .routers import login, users
 from .settings import settings
@@ -46,7 +45,6 @@ app.include_router(login.router)
 @app.on_event("startup")
 async def init_db() -> None:
     await setup_logging()
-    await create_super_user()
 
 
 @app.get("/")
