+++ logstash/central.conf
@@ -10,6 +10,11 @@ input {
       remove_field => [ "@timestamp", "@version" , "created" , "command", "version"]
       rename => {"msg" => "message"}
     }
+    mutate {
+        convert => {
+            "level" => "string"
+        }
+    }
 } output {
   elasticsearch {
     hosts => ["elasticsearch:9200"]
